Split samples into multiple files, update README.md
