+++ sdk/translation/azure-ai-translation-text/samples/text_translation_break_sentence.py
@@ -1,0 +1,74 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+# --------------------------------------------------------------------------
+
+"""
+FILE: text_translation_break_sentence.py
+
+DESCRIPTION:
+    This file contains sample snippets for the Text Translation service.
+
+USAGE:
+    python text_translation_break_sentence.py
+
+    Set the text translation endpoint environment variables with your own value before running the samples:
+        AZURE_TEXT_TRANSLATION_ENDPOINT - the endpoint to your Text Translation resource.
+    Note: the endpoint must be formatted to use the custom domain name for your resource:
+        https:\\<NAME-OF-YOUR-RESOURCE>.cognitiveservices.azure.com\
+"""
+
+from azure.core.exceptions import HttpResponseError
+from azure.ai.translation.text.models import (InputTextItem)
+
+# -------------------------------------------------------------------------
+# Text translation client
+# -------------------------------------------------------------------------
+import text_translation_client
+text_translator = text_translation_client.create_text_translation_client()
+
+def get_text_sentence_boundaries():
+    # [START get_text_sentence_boundaries]
+    try:
+        source_language = "zh-Hans"
+        source_script = "Latn"
+        input_text_elements = [ InputTextItem(text = "zhè shì gè cè shì。") ]
+
+        response = text_translator.find_sentence_boundaries(content = input_text_elements, language = source_language, script = source_script)
+        sentence_boundaries = response[0] if response else None
+
+        if sentence_boundaries:
+            detected_language = sentence_boundaries.detected_language
+            if detected_language:
+                print(f"Detected languages of the input text: {detected_language.language} with score: {detected_language.score}.")
+            print(f"The detected sentence boundaries:")
+            for boundary in sentence_boundaries.sent_len:
+                print(boundary)
+
+    except HttpResponseError as exception:
+        print(f"Error Code: {exception.error.code}")
+        print(f"Message: {exception.error.message}")
+    # [END get_text_sentence_boundaries]
+    
+def get_text_sentence_boundaries_auto():
+    # [START get_text_sentence_boundaries_auto]
+    try:
+        input_text_elements = [ InputTextItem(text = "This is a test. This is the second sentence.") ]
+
+        response = text_translator.find_sentence_boundaries(content = input_text_elements)
+        sentence_boundaries = response[0] if response else None
+
+        if sentence_boundaries:
+            detected_language = sentence_boundaries.detected_language
+            if detected_language:
+                print(f"Detected languages of the input text: {detected_language.language} with score: {detected_language.score}.")
+            print(f"The detected sentence boundaries:")
+            for boundary in sentence_boundaries.sent_len:
+                print(boundary)
+
+    except HttpResponseError as exception:
+        print(f"Error Code: {exception.error.code}")
+        print(f"Message: {exception.error.message}")
+    # [END get_text_sentence_boundaries_auto]
+
