+++ sdk/translation/azure-ai-translation-text/samples/text_translation_dictionary_lookup.py
@@ -1,0 +1,51 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+# --------------------------------------------------------------------------
+
+"""
+FILE: text_translation_dictionary_lookup.py
+
+DESCRIPTION:
+    This file contains sample snippets for the Text Translation service.
+
+USAGE:
+    python text_translation_dictionary_lookup.py
+
+    Set the text translation endpoint environment variables with your own value before running the samples:
+        AZURE_TEXT_TRANSLATION_ENDPOINT - the endpoint to your Text Translation resource.
+    Note: the endpoint must be formatted to use the custom domain name for your resource:
+        https:\\<NAME-OF-YOUR-RESOURCE>.cognitiveservices.azure.com\
+"""
+
+from azure.core.exceptions import HttpResponseError
+from azure.ai.translation.text.models import (InputTextItem)
+
+# -------------------------------------------------------------------------
+# Text translation client
+# -------------------------------------------------------------------------
+import text_translation_client
+text_translator = text_translation_client.create_text_translation_client()
+
+# -------------------------------------------------------------------------
+# Dictionary Lookup
+# -------------------------------------------------------------------------
+def get_text_dictionary_lookup():
+    # [START get_text_dictionary_lookup]
+    try:
+        source_language = "en"
+        target_language = "es"
+        input_text_elements = [ InputTextItem(text = "fly") ]
+
+        response = text_translator.lookup_dictionary_entries(content = input_text_elements, from_parameter = source_language, to = target_language)
+        dictionary_entry = response[0] if response else None
+
+        if dictionary_entry:
+            print(f"For the given input {len(dictionary_entry.translations)} entries were found in the dictionary.")
+            print(f"First entry: '{dictionary_entry.translations[0].display_target}', confidence: {dictionary_entry.translations[0].confidence}.")
+
+    except HttpResponseError as exception:
+        print(f"Error Code: {exception.error.code}")
+        print(f"Message: {exception.error.message}")
+    # [END get_text__dictionary_lookup]
