+++ sdk/translation/azure-ai-translation-text/samples/text_translation_translate.py
@@ -5,13 +5,13 @@
 # --------------------------------------------------------------------------
 
 """
-FILE: sample_snippets.py
+FILE: text_translation_translate.py
 
 DESCRIPTION:
     This file contains sample snippets for the Text Translation service.
 
 USAGE:
-    python sample_snippets.py
+    python text_translation_translate.py
 
     Set the text translation endpoint environment variables with your own value before running the samples:
         AZURE_TEXT_TRANSLATION_ENDPOINT - the endpoint to your Text Translation resource.
@@ -19,111 +19,17 @@ USAGE:
         https:\\<NAME-OF-YOUR-RESOURCE>.cognitiveservices.azure.com\
 """
 
-import os
-import text_translation_client
 from azure.core.exceptions import HttpResponseError
 from azure.ai.translation.text.models import (InputTextItem, TextType, ProfanityAction, ProfanityMarker)
 
 # -------------------------------------------------------------------------
 # Text translation client
 # -------------------------------------------------------------------------
-
+import text_translation_client
 text_translator = text_translation_client.create_text_translation_client()
 
 # -------------------------------------------------------------------------
-# Sample1_GetLanguages.md
-# -------------------------------------------------------------------------
-def get_text_translation_languages():
-    # [START get_text_translation_languages]
-    try:
-        response = response = text_translator.get_languages()
-
-        print(f"Number of supported languages for translate operation: {len(response.translation) if response.translation is not None else 0}")
-        print(f"Number of supported languages for transliterate operation: {len(response.transliteration) if response.transliteration is not None else 0}")
-        print(f"Number of supported languages for dictionary operations: {len(response.dictionary) if response.dictionary is not None else 0}")
-
-        if response.translation is not None:
-            print("Translation Languages:")
-            for key, value in response.translation.items():
-                print(f"{key} -- name: {value.name} ({value.native_name})")
-
-        if response.transliteration is not None:
-            print("Transliteration Languages:")
-            for key, value in response.transliteration.items():
-                print(f"{key} -- name: {value.name}, supported script count: {len(value.scripts)}")
-
-        if response.dictionary is not None:
-            print("Dictionary Languages:")
-            for key, value in response.dictionary.items():
-                print(f"{key} -- name: {value.name}, supported target languages count: {len(value.translations)}")
-
-    except HttpResponseError as exception:
-        print(f"Error Code: {exception.error.code}")
-        print(f"Message: {exception.error.message}")
-    # [END get_text_translation_languages]
-
-def get_text_translation_languages_scope():
-    # [START get_text_translation_languages_scope]
-    try:
-        scope = "translation"
-        response = response = text_translator.get_languages(scope=scope)
-
-        print(f"Number of supported languages for translate operation: {len(response.translation) if response.translation is not None else 0}")
-        print(f"Number of supported languages for transliterate operation: {len(response.transliteration) if response.transliteration is not None else 0}")
-        print(f"Number of supported languages for dictionary operations: {len(response.dictionary) if response.dictionary is not None else 0}")
-
-        if response.translation is not None:
-            print("Translation Languages:")
-            for key, value in response.translation.items():
-                print(f"{key} -- name: {value.name} ({value.native_name})")
-
-        if response.transliteration is not None:
-            print("Transliteration Languages:")
-            for key, value in response.transliteration.items():
-                print(f"{key} -- name: {value.name}, supported script count: {len(value.scripts)}")
-
-        if response.dictionary is not None:
-            print("Dictionary Languages:")
-            for key, value in response.dictionary.items():
-                print(f"{key} -- name: {value.name}, supported target languages count: {len(value.translations)}")
-
-    except HttpResponseError as exception:
-        print(f"Error Code: {exception.error.code}")
-        print(f"Message: {exception.error.message}")
-    # [END get_text_translation_languages_scope]
-    
-def get_text_translation_languages_culture():
-    # [START get_text_translation_languages_culture]
-    try:
-        accept_language = "es"
-        response = response = text_translator.get_languages(accept_language=accept_language)
-
-        print(f"Number of supported languages for translate operation: {len(response.translation) if response.translation is not None else 0}")
-        print(f"Number of supported languages for transliterate operation: {len(response.transliteration) if response.transliteration is not None else 0}")
-        print(f"Number of supported languages for dictionary operations: {len(response.dictionary) if response.dictionary is not None else 0}")
-
-        if response.translation is not None:
-            print("Translation Languages:")
-            for key, value in response.translation.items():
-                print(f"{key} -- name: {value.name} ({value.native_name})")
-
-        if response.transliteration is not None:
-            print("Transliteration Languages:")
-            for key, value in response.transliteration.items():
-                print(f"{key} -- name: {value.name}, supported script count: {len(value.scripts)}")
-
-        if response.dictionary is not None:
-            print("Dictionary Languages:")
-            for key, value in response.dictionary.items():
-                print(f"{key} -- name: {value.name}, supported target languages count: {len(value.translations)}")
-
-    except HttpResponseError as exception:
-        print(f"Error Code: {exception.error.code}")
-        print(f"Message: {exception.error.message}")
-    # [END get_text_translation_languages_culture]
-
-# -------------------------------------------------------------------------
-# Sample1_GetLanguages.md
+# Get text translation
 # -------------------------------------------------------------------------
 def get_text_translation():
     # [START get_text_translation]
@@ -174,7 +80,7 @@ def get_text_translation_with_transliteration():
 
 def get_text_translation_multiple_inputs():
     # [START get_text_translation_multiple_inputs]
-    try
+    try:
         target_languages = ["cs"]
         input_text_elements = [ 
             InputTextItem(text = "This is a test."),
@@ -367,115 +273,3 @@ def get_text_translation_custom():
         print(f"Message: {exception.error.message}")
     # [END get_text_translation_custom]
     
-# -------------------------------------------------------------------------
-# Sample3_Transliterate.md
-# -------------------------------------------------------------------------
-def get_text_transliteration():
-    # [START get_text_transliteration]
-    try:
-        language = "zh-Hans"
-        from_script = "Hans"
-        to_script = "Latn"
-        input_text_elements = [ InputTextItem(text = "这是个测试。") ]
-
-        response = text_translator.transliterate(content = input_text_elements, language = language, from_script = from_script, to_script = to_script)
-        transliteration = response[0] if response else None
-
-        if transliteration:
-            print(f"Input text was transliterated to '{transliteration.script}' script. Transliterated text: '{transliteration.text}'.")
-
-    except HttpResponseError as exception:
-        print(f"Error Code: {exception.error.code}")
-        print(f"Message: {exception.error.message}")
-    # [END get_text_transliteration]
-    
-# -------------------------------------------------------------------------
-# Sample4_BreakSentence.md
-# -------------------------------------------------------------------------    
-def get_text_sentence_boundaries():
-    # [START get_text_sentence_boundaries]
-    try:
-        source_language = "zh-Hans"
-        source_script = "Latn"
-        input_text_elements = [ InputTextItem(text = "zhè shì gè cè shì。") ]
-
-        response = text_translator.find_sentence_boundaries(content = input_text_elements, language = source_language, script = source_script)
-        sentence_boundaries = response[0] if response else None
-
-        if sentence_boundaries:
-            detected_language = sentence_boundaries.detected_language
-            if detected_language:
-                print(f"Detected languages of the input text: {detected_language.language} with score: {detected_language.score}.")
-            print(f"The detected sentence boundaries:")
-            for boundary in sentence_boundaries.sent_len:
-                print(boundary)
-
-    except HttpResponseError as exception:
-        print(f"Error Code: {exception.error.code}")
-        print(f"Message: {exception.error.message}")
-    # [END get_text_sentence_boundaries]
-    
-def get_text_sentence_boundaries_auto():
-    # [START get_text_sentence_boundaries_auto]
-    try:
-        input_text_elements = [ InputTextItem(text = "This is a test. This is the second sentence.") ]
-
-        response = text_translator.find_sentence_boundaries(content = input_text_elements)
-        sentence_boundaries = response[0] if response else None
-
-        if sentence_boundaries:
-            detected_language = sentence_boundaries.detected_language
-            if detected_language:
-                print(f"Detected languages of the input text: {detected_language.language} with score: {detected_language.score}.")
-            print(f"The detected sentence boundaries:")
-            for boundary in sentence_boundaries.sent_len:
-                print(boundary)
-
-    except HttpResponseError as exception:
-        print(f"Error Code: {exception.error.code}")
-        print(f"Message: {exception.error.message}")
-    # [END get_text_sentence_boundaries_auto]
-    
-# -------------------------------------------------------------------------
-# Sample5_DictionaryLookup.md
-# -------------------------------------------------------------------------
-def get_text_dictionary_lookup():
-    # [START get_text_dictionary_lookup]
-    try:
-        source_language = "en"
-        target_language = "es"
-        input_text_elements = [ InputTextItem(text = "fly") ]
-
-        response = text_translator.lookup_dictionary_entries(content = input_text_elements, from_parameter = source_language, to = target_language)
-        dictionary_entry = response[0] if response else None
-
-        if dictionary_entry:
-            print(f"For the given input {len(dictionary_entry.translations)} entries were found in the dictionary.")
-            print(f"First entry: '{dictionary_entry.translations[0].display_target}', confidence: {dictionary_entry.translations[0].confidence}.")
-
-    except HttpResponseError as exception:
-        print(f"Error Code: {exception.error.code}")
-        print(f"Message: {exception.error.message}")
-    # [END get_text_translation_dictionary_lookup]
-
-# -------------------------------------------------------------------------
-# Sample6_DictionaryExamples.md
-# -------------------------------------------------------------------------
-def get_text_dictionary_entries():
-    # [START get_text_dictionary_entries]
-    try:
-        source_language = "en"
-        target_language = "es"
-        input_text_elements = [ InputTextItem(text = "fly") ]
-
-        response = text_translator.lookup_dictionary_entries(content = input_text_elements, from_parameter = source_language, to = target_language)
-        dictionary_entry = response[0] if response else None
-
-        if dictionary_entry:
-            print(f"For the given input {len(dictionary_entry.translations)} entries were found in the dictionary.")
-            print(f"First entry: '{dictionary_entry.translations[0].display_target}', confidence: {dictionary_entry.translations[0].confidence}.")
-
-    except HttpResponseError as exception:
-        print(f"Error Code: {exception.error.code}")
-        print(f"Message: {exception.error.message}")
-    # [END get_text_dictionary_entries]
