+++ sdk/translation/azure-ai-translation-text/samples/text_translation_transliterate.py
@@ -1,0 +1,51 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+# --------------------------------------------------------------------------
+
+"""
+FILE: text_translation_transliterate.py
+
+DESCRIPTION:
+    This file contains sample snippets for the Text Translation service.
+
+USAGE:
+    python text_translation_transliterate.py
+
+    Set the text translation endpoint environment variables with your own value before running the samples:
+        AZURE_TEXT_TRANSLATION_ENDPOINT - the endpoint to your Text Translation resource.
+    Note: the endpoint must be formatted to use the custom domain name for your resource:
+        https:\\<NAME-OF-YOUR-RESOURCE>.cognitiveservices.azure.com\
+"""
+
+from azure.core.exceptions import HttpResponseError
+from azure.ai.translation.text.models import (InputTextItem)
+
+# -------------------------------------------------------------------------
+# Text translation client
+# -------------------------------------------------------------------------
+import text_translation_client
+text_translator = text_translation_client.create_text_translation_client()
+
+# -------------------------------------------------------------------------
+# Text translation transliteration
+# -------------------------------------------------------------------------
+def get_text_transliteration():
+    # [START get_text_transliteration]
+    try:
+        language = "zh-Hans"
+        from_script = "Hans"
+        to_script = "Latn"
+        input_text_elements = [ InputTextItem(text = "这是个测试。") ]
+
+        response = text_translator.transliterate(content = input_text_elements, language = language, from_script = from_script, to_script = to_script)
+        transliteration = response[0] if response else None
+
+        if transliteration:
+            print(f"Input text was transliterated to '{transliteration.script}' script. Transliterated text: '{transliteration.text}'.")
+
+    except HttpResponseError as exception:
+        print(f"Error Code: {exception.error.code}")
+        print(f"Message: {exception.error.message}")
+    # [END get_text_transliteration]
